Changed logic for adding products to shopping cart
I removed the select amount from products/index.html.erb
So now user can only add specific product to cart and edit the amount
directly in the shopping cart.
So the logic for buying only that many products that are available will be on the
cart_controller side.
The specs also have changed due to change in parameters sent form index.html.erb