Added query object for finding bestseller and service objects
I divided the flow so the index method inside admins/statistics_controller
calls the service objects for bestseller and sold_products_amount which
call the query object for both methods.

I added also low level cache on the service objects, so bestseller is cached
for 5 days and the sold_product_amount is watching for changes on PurchaseProduct
I did it so because it seems that this query on sold_product_amount will be called
too often but it is still less often than when it would be on BessellerQuery#initialize
where the select in run to find all purchase products that belong to "bought" purchase.