Added calculation pattern for statistics
I changed the calculations for statistics.
Instead holding it separately for users categories etc
i put them all in one file and reutrned as hash in controller,
so now in the index view i can access them with ease.
I changed also the cache duration and removed the cache_key.
It seems to be unnecessary to calculate the result each time when
some record is added to database. I assumed that it will be enough to
refresh those data once for every 7 days.